@@ -4,7 +4,7 @@ import (
 	"bytes"
 	"encoding/json"
 	"html/template"
-	"io/ioutil"
+	"io"
 	"log"
 	"net/http"
 	"os"
@@ -21,9 +21,11 @@ const (
 
 type contact struct {
 	FullName    string
+	Company     string
 	Email       string
 	ContactType string
 	Message     string
+	APIKey      string
 }
 
 func addresses(a string) (addr []*string) {
@@ -51,15 +53,108 @@ func contactMailer(w http.ResponseWriter, r *http.Request) {
 		return
 	}
 
-	if err = contactNotification(contact{
-		FullName:    r.FormValue("fullname"),
-		Email:       r.FormValue("email"),
-		ContactType: r.FormValue("contact_type"),
+	fullName := r.FormValue("fullname")
+	company := r.FormValue("company")
+	email := r.FormValue("email")
+	apiKey := ""
+	contactType := r.FormValue("contact_type")
+	if contactType == "sdk-liquid-api-key" {
+		apiKey, err = generateKey(fullName, email, company)
+		if err != nil {
+			log.Printf("Error in generateKey(\"%v\",\"%v\",\"%v\"): %v", fullName, email, company, err)
+		}
+	}
+
+	c := contact{
+		FullName:    fullName,
+		Company:     company,
+		Email:       email,
+		ContactType: contactType,
 		Message:     r.FormValue("message"),
-	}); err != nil {
+		APIKey:      apiKey,
+	}
+	if err = contactNotification(c); err != nil {
 		http.Error(w, "", http.StatusBadRequest)
 		return
 	}
+
+	if apiKey != "" {
+		sendAPIKey(c)
+	}
+}
+
+func sendAPIKey(c contact) error {
+	var html bytes.Buffer
+
+	tpl := `
+Hello {{ .FullName }},<br /><br />
+Please use the following API key for the Breez SDK (Liquid implementation):<br /><br />
+<pre>{{ .APIKey }}</pre><br />
+If you have any questions, you can reply to this email.<br />
+<br />
+Best,<br />
+The Breez team
+	`
+	t, err := template.New("APIKeyEmail").Parse(tpl)
+	if err != nil {
+		return err
+	}
+
+	if err := t.Execute(&html, c); err != nil {
+		return err
+	}
+
+	sess, err := session.NewSession(&aws.Config{})
+	if err != nil {
+		log.Printf("Error in session.NewSession: %v", err)
+		return err
+	}
+	svc := ses.New(sess)
+
+	input := &ses.SendEmailInput{
+		Destination: &ses.Destination{
+			BccAddresses: addresses(os.Getenv("APIKEY_EMAIL_BCC")),
+			ToAddresses:  []*string{&c.Email},
+		},
+		Message: &ses.Message{
+			Body: &ses.Body{
+				Html: &ses.Content{
+					Charset: aws.String(charset),
+					Data:    aws.String(html.String()),
+				},
+			},
+			Subject: &ses.Content{
+				Charset: aws.String(charset),
+				Data:    aws.String("API key"),
+			},
+		},
+		Source: aws.String(os.Getenv("APIKEY_EMAIL_FROM")),
+	}
+	// Attempt to send the email.
+	result, err := svc.SendEmail(input)
+	if err != nil {
+		if aerr, ok := err.(awserr.Error); ok {
+			switch aerr.Code() {
+			case ses.ErrCodeMessageRejected:
+				log.Println(ses.ErrCodeMessageRejected, aerr.Error())
+			case ses.ErrCodeMailFromDomainNotVerifiedException:
+				log.Println(ses.ErrCodeMailFromDomainNotVerifiedException, aerr.Error())
+			case ses.ErrCodeConfigurationSetDoesNotExistException:
+				log.Println(ses.ErrCodeConfigurationSetDoesNotExistException, aerr.Error())
+			default:
+				log.Println(aerr.Error())
+			}
+		} else {
+			// Print the error, cast err to awserr.Error to get the Code and
+			// Message from an error.
+			log.Println(err.Error())
+		}
+		return err
+	}
+
+	log.Printf("Transmission sent with result:\n%v", result)
+
+	return nil
 }
 
 func contactNotification(c contact) error {
@@ -67,11 +162,17 @@ func contactNotification(c contact) error {
 	var html bytes.Buffer
 
 	tpl := `
-	<div>FullName: {{ .FullName }}</div>
+	<div>Full Name: {{ .FullName }}</div>
+	{{if .Company -}}
+	<div>Company: {{ .Company }}</div>
+	{{end -}}
 	<div>Email: {{ .Email }}</div>
 	<div>Contact Type: {{ .ContactType }}</div>
 	<div>Message:<br /> {{ .Message }}</div>
-	
+	{{if .APIKey -}}
+	<div>API Key: <pre>{{ .APIKey }}</pre></div>
+	{{end -}}
+
 	`
 	t, err := template.New("ContactEmail").Parse(tpl)
 	if err != nil {
@@ -144,7 +245,7 @@ func paymentFailure(w http.ResponseWriter, r *http.Request) {
 		http.Error(w, "404 not found.", http.StatusNotFound)
 		return
 	}
-	b, err := ioutil.ReadAll(r.Body)
+	b, err := io.ReadAll(r.Body)
 	defer r.Body.Close()
 	if err != nil {
 		http.Error(w, err.Error(), 500)
